chore: move File to codegen-core
