+++ codegen/scala-gen/src/main/scala/com/akkaserverless/codegen/scalasdk/impl/ActionTestKitGenerator.scala
@@ -16,7 +16,7 @@
 
 package com.akkaserverless.codegen.scalasdk.impl
 
-import com.akkaserverless.codegen.scalasdk.File
+import com.lightbend.akkasls.codegen.File
 import com.lightbend.akkasls.codegen.Format
 import com.lightbend.akkasls.codegen.Imports
 import com.lightbend.akkasls.codegen.ModelBuilder
